Create Branding procedures
Since Owncloud logo's are not a simple graphics editing exercise it is important to understand specifics regarding the editing of the files required to change the logo's or branding of your owncloud installation. These easy to follow steps provide every step necessary to successfully re-brand your owncloud. After all, it's called owncloud so, shy not make it your own.