1- Many claim are erroneous like the existance of a folder default in themes/ 2- All default files are in /core and /settings and the current version doesn't use the theming system for its default interface.
There were no information about what file should be updated or are of interest.

Please someone validate that claim, I'm not sure if I'm right that this file cannot be located in theme folder to be effective!

If I'm right, please open a bug in Core to add a feature to modify these value from admin page.

- /owncloud/lib/private/defaults.php # Contain a configuration section to rename the application to your custom name.  !! This file cannot be copied in your custom theme folder.