Added Required to documentation, website, repository, bugs
These sections should be **Required** as they always should exist (and with github generally automatically do). For the documentation there should be at least a stub provided, for others to join in. Therefore they should be mentioned to make it easier and faster for contributers to find the right ones.

Sorry if I might not have followed the proper way to propose this change, I just did not find a proper and obvious way to start the discussion. Therefore I thought it might be easier and faster to propose the change and then the consideration can start. As for this it was only a small thing to do, this was a reasonable choice for me. If this does not work for you please help me to find the proper way.