Small restructure of the documentation directory tree
The intent of this change is to begin to make the documentation tree
that much simpler to navigate. This change, specifically, moves
admin_manual, developer_manual, and user_manual in to the structure:

```
manual/
  admin
  developer
  user
```

The respective `conf.py` scripts have been updated to reflect the new
structure as well. The developer and user manuals still build, however the
admin manual doesn't. I'm not sure if that was the case previously.