Reworked the Building section
This section appeared to have a lot of redundancy, which could be
compacted, making the section easier to work with. Given that, the
section's been split in to two sections: "Installing The Required
Dependencies" & "Building".

I also reworked the installation commands for the dependencies on the
Linux platforms, such that they can be copied and pasted into a terminal
and run.