Fix bash error that will cause script to die
Using `set -e` means that  every line in the script is expected to
succeed in its own right. Running something just to try it and testing
the return code later won't work in this mode. If you want to branch
based on the return code of a command (or in this case pipeline) you
need to make it part of the if statement so bash has somewhere to go
while still succeeding after every command structure.