Revise the OCS Share Api Developer Documentation
This commit adds a host of code examples for interacting with the API,
documented in the OCS Share API section of the documentation. It
specifically adds examples for most of the endpoints, in Curl, PHP,
Ruby, and Go. It also adds response payload examples.

In amongst this commit, the structure of that section of the
documentation was also updated to give it a better sense of structure
and style.