Make the system requirements easier to understand
While you can never force someone to read something, and never be sure
that they have, by making it as clear as possible, I'd suggest that you
can greatly increase the chances that they will read it. Given that,
this commit is being submitted to improve the readability of this
section.

This is implemented to support #2784.