Refactor tags and permissions data into tables
This is part of addressing #2729. The reason for doing so is to make the
information much more readable. When a lot of information, such as this,
is presented in an unstructured way, it becomes difficult to read. Given
that, a table format was chosen to make it a lot simpler.