Heavily re-worked the ext_storage file
Strictly speaking not all of the changes in this commit were necessary.
However, to ensure that the text is as readable as possible, that it
flows from beginning to end, etc, the rework was carried out as part of
this ticket.

I don't mean this in a mean, nor condescending way. But, on reading
through it, the text wasn't as clear to me as I felt it should be. So it
seemed worthwhile to rework it somewhat.