Update the Ansible provisioner to support vhosts for the manuals
The reason behind this is to make the org version of the docs available
locally, so that it's as simple as possible to test changes to the
documentation before a commit is made or a PR is merged. The prime
motivation for this is because the available tools to lint
reStructuredText files, when using the Sphinx-doc additions is limited.
So, this appears to be the next best thing.