Caution against running some cron jobs in parallel on multiple servers
In https://github.com/owncloud/enterprise/issues/1882 the request was
made to caution users against running ExpireTrash and ExpireVersions in
parallel on multiple servers as there's insufficient locking at the
moment. So, for that reason, an admonition was added in this commit to
make it clear to users that this is the case.

This fixes owncloud/enterprise/issues/1882.