Simplified the directory structure of the configuration docs
There's a couple of reasons for doing this, but the main one is that it
could be simplified, resulting in an easier to read section of the docs.
Now that it's done, the main table of contents can also be compressed
further, making each level of the documentation be more concise than
before.