Change style of active navigation link
james.rhea from the Write the Docs Slack channel pointed out that the
active link in the navigation menu is always white — on a white
background. This commit was created to fix that readability problem as
well as to provide a better UI experience for the user, making the
content that much more intuitive to read and navigate.