Update "users/getuser" XML output example
The information in the existing example XML output, while not incorrect,
was incomplete. It only included values for email, whether the user was
enabled, and very minimalist quota representation.

The response actually includes displayname, home directory, two factor
auth enabled status, and richer quota information.  So the example XML
output was updated to include this.