Update unit testing further reading links
Some of the links were just links, and would be more meaningful if the
user knew what the link was before they clicked on it. And the final
item in the list was only a name, without a link. Better to link it and
save the user the trouble of having to google it.