Document that the installation wizard should not be unprotected
If the installation wizard is left unprotected, there is the possibility
that a malicious actor could compromise the installation — this is a
known problem, made worse by certificate transparency
(https://www.certificate-transparency.org). To help mitigate against the
potential threat, it's important to inform the user that this could
happen and some options that they can take to help mitigate it from
happening.