Fix the rendering of the post installation script
It wasn't rendering because of a trailing double-colon just a little
above the inclusion of the script. Secondly, to make the code example
that bit easier to manage, the code's now being included from a separate
code file. This has the extra benefit of allowing the code to be linted,
etc. It fixes #3820.