Update the manual upgrade documentation
Some content was integrated from the upgrade index page, and other
content was cleaned up. The cleanup includes removing outdated content
(such as content relating to older versions, which is, in-effect, no
longer applicable to current versions). Headers were restructured and
renamed so that they were clearer. Content was reorganised to make more
sense and to be clearer and more succinct.