Add clear and concise instructions for installing Sphinx-Doc
This has been a source of much confusion and frustration and it’s not
properly documented by the official documentation, to the best of my
knowledge. Given that, I want to make it clear here. Ok, it saves me the
time and effort of finding the instructions again if Vagrant dies on me.