Major rework of the user provisioning API documentation
After looking at the documentation, it was clear that it needed quite a
bit of love to bring it up to the current standard. This commit changes
headers, makes the endpoints and their arguments clearer, by storing
them in tables, converts the existing examples to curl console examples,
and a few other changes. This relates to #4079.