Fix the theme to render the footer links correctly
This commit fixes the broken footer layout, so that the three columns of
links are rendered correctly, whether when using a desktop or
responsive/mobile browser. On the desktop browser, they'll render beside
each other, taking up about a third of the available width each. When
viewed from a responsive browser, they'll render centered, one under
each other.  The attempt is to provide a Bootstrap-esque 3-column
layout. It's not quite perfect, but close enough for now.