Move create shares programmatically section to the file sharing config section
Having the create shares programmatically section as part of the Windows
Network Drive (WND) section doesn't make sense, as the information's
hidden away in an area that only a small percentage of users will ever
read. What's more, it makes sense to put it with a section that it's a
part of. That's why it's being moved.